+++ Source/Project64/UserInterface/Debugger/ScriptSystem.cpp
@@ -202,7 +202,7 @@ void CScriptSystem::UpdateCpuCbListInfo(volatile JSCpuCbListInfo & info, JSAppCa
 {
     uint32_t minAddrStart = 0;
     uint32_t maxAddrEnd = 0;
-    int numCacheEntries = 0;
+    size_t numCacheEntries = 0;
     bool bCacheExceeded = false;
 
     for (JSAppCallback & callback : callbacks)
